refactor: remove unnecessary floating-point casts
The algorithm that generates the identity CLUT computes values using
f64's because the result is a fraction in the range [0.0, 1.0) and then
multiplies this to convert it to a u8. This can be simplified by
multiplying by 255 first and then using integer division.

This does not actually improve performance, I assume LLVM optimizes out
the floating-point math.